backend/outbreak_api/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
lightgbm
pydantic
google-cloud-firestore
numpy
twilio